refactor: change pyspark code using best practices
