+++ src/spark-processor/app/processor.py
@@ -1,11 +1,12 @@
-from pyspark.sql import SparkSession
-from pyspark.sql.types import StructType,StructField, StringType, IntegerType, BooleanType, FloatType, LongType
 import pyspark.sql.functions as sf
-
-import influxdb_client
-from influxdb_client.client.write_api import SYNCHRONOUS
-
 import os
+
+from pyspark.sql.types import StructType,StructField, StringType, IntegerType, BooleanType, FloatType, LongType
+
+# importando variáveis de ambiente globais
+from utils import KAFKA_TOPIC, KAFKA_BOOSTSTRAP_SERVERS, INFLUXDB_BUCKET, INFLUXDB_ORG, INFLUXDB_URL, INFLUXDB_TOKEN, MINIO_USER, MINIO_PASSWORD, MINIO_BUCKET, POSTGRES_USER, POSTGRES_PASSWORD, POSTGRES_DB, CATALOG_NAME, PROCESSING_TIME_WINDOW
+# importando funções auxiliares
+from utils import get_spark_session, write_influxdb, write_raw_to_iceberg, create_schema_iceberg_if_not_exist
 
 #{
 #    "stream":"btcusdt@aggTrade",
@@ -25,167 +26,15 @@ import os
 #   "m": true,        // Is the buyer the market maker?
 # }
 
-KAFKA_TOPIC = "binance-trades-raw"
-KAFKA_BOOSTSTRAP_SERVERS = "kafka:29092"
-INFLUXDB_BUCKET = "trades_raw"
-INFLUXDB_ORG = "crypto_pipeline_org"
-INFLUXDB_URL="http://influxdb:8086"
-
-INFLUXDB_TOKEN = os.getenv("INFLUXDB_TOKEN")
-MINIO_USER = os.getenv("AWS_ACCESS_KEY_ID")
-MINIO_PASSWORD = os.getenv("AWS_SECRET_ACCESS_KEY")
-POSTGRES_USER= os.getenv("POSTGRES_USER")
-POSTGRES_PASSWORD = os.getenv("POSTGRES_PASSWORD")
-
-def get_spark_session(postgres_user, 
-                      postgres_password,
-                      minio_user, 
-                      minio_password,
-                      catalog_name, 
-                      postgres_db="iceberg", 
-                      minio_bucket_name="cripto-data"):
-
-    spark = SparkSession \
-        .builder \
-        .appName("cryptoDataPipelineStreaming") \
-        .master("spark://spark-master:7077") \
-        .config("spark.sql.caseSensitive", "true") \
-        .config("spark.sql.extensions", "org.apache.iceberg.spark.extensions.IcebergSparkSessionExtensions") \
-        .config(f"spark.sql.catalog.{catalog_name}", "org.apache.iceberg.spark.SparkCatalog") \
-        .config(f"spark.sql.catalog.{catalog_name}.type", "jdbc") \
-        .config(f"spark.sql.catalog.{catalog_name}.warehouse", f"s3a://{minio_bucket_name}/") \
-        .config(f"spark.sql.catalog.{catalog_name}.uri", f"jdbc:postgresql://postgres:5432/{postgres_db}") \
-        .config(f"spark.sql.catalog.{catalog_name}.jdbc.verifyServerCertificate", "False") \
-        .config(f"spark.sql.catalog.{catalog_name}.jdbc.useSSL", "False") \
-        .config(f"spark.sql.catalog.{catalog_name}.jdbc.user", postgres_user) \
-        .config(f"spark.sql.catalog.{catalog_name}.jdbc.password", postgres_password) \
-        .config("spark.hadoop.fs.s3a.endpoint", "http://minio:9000") \
-        .config("spark.hadoop.fs.s3a.access.key", minio_user) \
-        .config("spark.hadoop.fs.s3a.secret.key", minio_password) \
-        .config("spark.hadoop.fs.s3a.path.style.access", "true") \
-        .config("spark.hadoop.fs.s3a.impl", "org.apache.hadoop.fs.s3a.S3AFileSystem") \
-        .getOrCreate()
-    
-    return spark
-
-def write_influxdb_not_optimizer(batch_df, batch_id):
-
-    # =============== HOT PATH =====================
-    print(f"--- Processando Lote ID: {batch_id} ---")
-    
-    client = influxdb_client.InfluxDBClient(
-        url=INFLUXDB_URL,
-        token=INFLUXDB_TOKEN,
-        org=INFLUXDB_ORG
-    )
-
-    write_api = client.write_api(write_options=SYNCHRONOUS)
-
-    df = batch_df.toPandas()
-
-    points = []
-    for index, row in df.iterrows():
-        p = influxdb_client.Point("trades_summary") \
-            .tag("symbol", row['symbol']) \
-            .field("total_quantity", row['total_quantity']) \
-            .field("average_price", row['average_price']) \
-            .time(row.window['start'])
-        
-        points.append(p)
-    
-    write_api.write(bucket=INFLUXDB_BUCKET, org=INFLUXDB_ORG, record=points)
-    
-    print(f"Lote {batch_id} com {len(df)} linhas escrito no InfluxDB.")
-    
-    client.close()
-
-def write_influxdb(iterator_of_rows):
-    client = influxdb_client.InfluxDBClient(
-        url=INFLUXDB_URL,
-        token=INFLUXDB_TOKEN,
-        org=INFLUXDB_ORG
-    )
-
-    write_api = client.write_api(write_options=SYNCHRONOUS)
-
-    points = []
-    for row in iterator_of_rows:
-        p = influxdb_client.Point("trades_summary") \
-            .tag("symbol", row['symbol']) \
-            .field("total_quantity", row['total_quantity']) \
-            .field("average_price", row['average_price']) \
-            .time(row.window['start'])
-        
-        points.append(p)
-    
-    write_api.write(bucket=INFLUXDB_BUCKET, org=INFLUXDB_ORG, record=points)
-
-    client.close()
-    
-def write_raw_to_iceberg(batch_df, batch_id):
-      # =============== COLD PATH =====================
-    try:
-        print(f"--- Escrevendo lote {batch_id} no Iceberg (Cold Path) ---")
-        
-        batch_df = batch_df.withColumn("trade_date", sf.to_date(sf.col("event_timestamp")))
-
-        table_name = "cripto_catalog.raw.btc_trades_agg"
-
-        if batch_df.sparkSession.catalog.tableExists(table_name):
-            batch_df.writeTo(table_name) \
-            .partitionedBy("trade_date") \
-            .append()
-        else:
-            batch_df.writeTo(table_name) \
-            .partitionedBy("trade_date") \
-            .create()
-        
-        
-        print(f"--- Lote {batch_id} escrito no Iceberg com sucesso. ---")
-    
-    except Exception as e:
-        print(f"Erro ao escrever no Iceberg: {e}")
-
-    # Libera o DataFrame da memória
-    batch_df.unpersist()
-
-def write_raw_to_minio(batch_df, batch_id,
-                      minio_bucket_name = "cripto-data"):
-    """
-    Salva um micro-lote de dados brutos como arquivos Parquet no MinIO.
-    """
-    print(f"--- Escrevendo lote BRUTO {batch_id} como Parquet no MinIO ---")
-    try:
-        # Adiciona a coluna de data para particionar as pastas
-        df_for_writing = batch_df.withColumn("trade_date", sf.to_date(sf.col("event_timestamp")))
-        
-        # Define o caminho de destino no MinIO
-        output_path = f"s3a://{minio_bucket_name}/streaming/btc/"
-
-        # Escreve o DataFrame no formato Parquet
-        df_for_writing.write \
-            .mode("append") \
-            .partitionBy("trade_date") \
-            .parquet(output_path)
-        
-        print(f"--- Lote BRUTO {batch_id} escrito com sucesso em '{output_path}' ---")
-    
-    except Exception as e:
-        print(f"!!! Erro ao escrever no MinIO: {e}")
 
 def main():
-    catalog_name="cripto_catalog"
-    postgres_db="iceberg"
-    minio_bucket_name = "cripto-data"
-    processing_time_window = '15 seconds'
-
     spark = get_spark_session(postgres_user = POSTGRES_USER,
                             postgres_password = POSTGRES_PASSWORD,
                             minio_user=MINIO_USER,
                             minio_password=MINIO_PASSWORD,
-                            catalog_name=catalog_name, 
-                            postgres_db=postgres_db,
-                            minio_bucket_name = minio_bucket_name)
+                            catalog_name=CATALOG_NAME, 
+                            postgres_db=POSTGRES_DB,
+                            minio_bucket_name = MINIO_BUCKET)
 
     user_schema = StructType([ \
         StructField("e", StringType(), True), \
@@ -200,12 +49,14 @@ def main():
         StructField("m", BooleanType(), True)
       ])
 
+    create_schema_iceberg_if_not_exist(spark, catalog_name= CATALOG_NAME, schema_name="raw")
+
     # defining source
     binance_df_raw = spark \
                     .readStream \
                     .format("kafka") \
                     .option("kafka.bootstrap.servers", KAFKA_BOOSTSTRAP_SERVERS) \
-                    .option("subscribe", KAFKA_TOPIC) \
+                    .option("subscribePattern", "trades-.*") \
                     .load()
 
     # 1. Cast para String
@@ -241,15 +92,15 @@ def main():
     query_raw = df_with_timestamp.writeStream \
         .outputMode("append") \
         .foreachBatch(write_raw_to_iceberg) \
-        .option("checkpointLocation", f"s3a://{minio_bucket_name}/spark_checkpoints/cold_path_sink") \
-        .trigger(processingTime=processing_time_window) \
+        .option("checkpointLocation", f"s3a://{MINIO_BUCKET}/spark_checkpoints/cold_path_sink") \
+        .trigger(processingTime=PROCESSING_TIME_WINDOW) \
         .start()
 
 
     # 6. Aplicando window function e agregações
-    df_windowed = df_with_timestamp.withWatermark("event_timestamp", "15 seconds").groupBy(
+    df_windowed = df_with_timestamp.withWatermark("event_timestamp", "60 seconds").groupBy(
         sf.col("symbol"),
-        sf.window(sf.col("event_timestamp"), "10 seconds")
+        sf.window(sf.col("event_timestamp"), "30 seconds")
     ).agg(
         sf.sum("quantity").alias("total_quantity"), 
         sf.avg("price").alias("average_price")
@@ -259,8 +110,8 @@ def main():
     query_aggregated = df_windowed.writeStream \
         .outputMode("update") \
         .foreachBatch(lambda batch_df, batch_id: batch_df.foreachPartition(write_influxdb)) \
-        .option("checkpointLocation", f"s3a://{minio_bucket_name}/spark_checkpoints/influxdb_sink") \
-        .trigger(processingTime=processing_time_window) \
+        .option("checkpointLocation", f"s3a://{MINIO_BUCKET}/spark_checkpoints/influxdb_sink") \
+        .trigger(processingTime=PROCESSING_TIME_WINDOW) \
         .start()
